requirements.txt
streamlit
pandas
cassandra-driver
scikit-learn
matplotlib
seaborn
numpy